% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{exampleSettings}
\alias{exampleSettings}
\title{Incremental Read Only Settings}
\format{
Named list of settings.
}
\usage{
exampleSettings
}
\description{
Incremental Read Only Settings
}
\keyword{datasets}
